parte cliente operativa

+++ src/main/java/com/ejemplo/tienda_online/controller/DataInitController.java
@@ -7,7 +7,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -29,31 +28,31 @@ public class DataInitController {
             Producto.builder()
                 .nombre("Laptop Gaming")
                 .descripcion("Laptop para gaming de alta gama con procesador Intel i7")
-                .precio(BigDecimal.valueOf(1299.99))
+                .precio(1299.99)
                 .stock(10)
                 .build(),
             Producto.builder()
                 .nombre("Mouse Inalámbrico")
                 .descripcion("Mouse inalámbrico ergonómico con sensor óptico")
-                .precio(BigDecimal.valueOf(29.99))
+                .precio(29.99)
                 .stock(50)
                 .build(),
             Producto.builder()
                 .nombre("Teclado Mecánico")
                 .descripcion("Teclado mecánico con retroiluminación RGB")
-                .precio(BigDecimal.valueOf(89.99))
+                .precio(89.99)
                 .stock(25)
                 .build(),
             Producto.builder()
                 .nombre("Monitor 4K")
                 .descripcion("Monitor 4K de 27 pulgadas con tecnología IPS")
-                .precio(BigDecimal.valueOf(399.99))
+                .precio(399.99)
                 .stock(15)
                 .build(),
             Producto.builder()
                 .nombre("Auriculares Gaming")
                 .descripcion("Auriculares gaming con sonido surround 7.1")
-                .precio(BigDecimal.valueOf(159.99))
+                .precio(159.99)
                 .stock(30)
                 .build()
         );
